tool-suite/cluster-settings/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioice==0.9.0
aioresponses==0.7.6
aiortc==1.9.0
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.2.post1
APScheduler==3.10.1
attrs==24.2.0
av==12.3.0
awscli==1.33.44
backoff==2.2.1
boto3==1.34.123
botocore==1.34.162
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
Cython==3.0.0
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docker==6.1.3
docutils==0.16
fastapi==0.110.3
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
GPUtil==1.4.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.1
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
imageio==2.36.0
inference==0.24.0
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.15.1
openai==1.52.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
pendulum==3.0.0
piexif==1.1.3
pillow==10.4.0
pluggy==1.5.0
prettytable==3.11.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.21.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
PuLP==2.9.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase64==1.4.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydot==3.0.2
pyee==12.0.0
Pygments==2.18.0
pylibsrtp==0.10.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
pytest==8.3.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.5.2
rpds-py==0.20.0
rsa==4.7.2
s3transfer==0.10.3
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
skypilot==0.5.0
sniffio==1.3.1
starlette==0.37.2
structlog==24.4.0
supervision==0.22.0
sympy==1.13.1
tabulate==0.9.0
tifffile==2024.9.20
time-machine==2.16.0
tldextract==5.1.2
tokenizers==0.20.1
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
triton==3.1.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.20
wcwidth==0.2.13
websocket-client==1.8.0
yarl==1.16.0
zxing-cpp==2.2.0
